+++ src/App.js
@@ -2,8 +2,7 @@ import logo from "./logo.svg";
 import "./App.css";
 import RecycleBin from "./RecycleBin";
 import { useState } from "react";
-import List from "./components/List/List";
-import AddTask from "./components/List/AddTask";
+import List from "./components/List";
 
 function App() {
   const [todos, setTodos] = useState(["write", "read", "sleep", "eat"]);
@@ -11,7 +10,6 @@ function App() {
     <div className="App">
       <header className="App-header">
         <List todos={todos} />
-        <AddTask />
       </header>
       <RecycleBin />
     </div>
