+++ src/components/AddTask.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-
 export default function AddTask() {
   return (
     <div>
